requirements.txt
# PDF Processing
pdfplumber>=0.10.0

# LLM & Embeddings
langchain-google-genai>=1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Vector Database
chromadb>=0.4.22

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Frontend
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0
requests>=2.31.0